requirements.txt
beautifulsoup4
Django
mysqlclient
python-dateutil
requests